added status codes to controller and modified method save

+++ src/main/java/com/example/bookstoreapp/service/impl/BookServiceImpl.java
@@ -2,13 +2,16 @@ package com.example.bookstoreapp.service.impl;
 
 import com.example.bookstoreapp.dto.BookDto;
 import com.example.bookstoreapp.dto.CreateBookRequestDto;
+import com.example.bookstoreapp.dto.UpdateBookDto;
 import com.example.bookstoreapp.entity.Book;
 import com.example.bookstoreapp.exception.EntityNotFoundException;
 import com.example.bookstoreapp.mapper.BookMapper;
 import com.example.bookstoreapp.repository.BookRepository;
 import com.example.bookstoreapp.service.BookService;
+
 import java.util.List;
 import java.util.Optional;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -17,6 +20,8 @@ import org.springframework.stereotype.Service;
 public class BookServiceImpl implements BookService {
 
     private static final String CAN_T_FIND_BOOK_MSG = "Can't find book";
+    public static final String T_UPDATE_BOOK_BY_ID_MSG = "Can't update book by id ";
+    public static final String CAN_T_DELETE_BOOK_BY_ID_MSG = "Can't delete book by id ";
 
     private final BookRepository bookRepository;
 
@@ -43,18 +48,26 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public BookDto updateBook(Long id, CreateBookRequestDto book) {
-        Optional<Book> bookById = bookRepository.findById(id);
-        if (bookById.isPresent()) {
-            bookRepository.updateBookById(id, book);
-        }
-        return bookMapper
-                .bookToBookDto(bookById.orElseThrow(()
-                        -> new EntityNotFoundException("Can't update book by id " + id)));
+    public BookDto updateBook(Long id, UpdateBookDto book) {
+        Book updatedBook = bookRepository.findById(id).orElseThrow(() ->
+                new EntityNotFoundException(T_UPDATE_BOOK_BY_ID_MSG + id));
+        updatedBook.setTitle(book.getTitle());
+        updatedBook.setAuthor(book.getAuthor());
+        updatedBook.setIsbn(book.getIsbn());
+        updatedBook.setPrice(book.getPrice());
+        updatedBook.setDescription(book.getDescription());
+        updatedBook.setCoverImage(book.getCoverImage());
+
+        bookRepository.save(updatedBook);
+        return bookMapper.bookToBookDto(updatedBook);
     }
 
     @Override
     public void deleteById(Long id) {
-        bookRepository.deleteById(id);
+        if (bookRepository.existsById(id)) {
+            bookRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException(CAN_T_DELETE_BOOK_BY_ID_MSG + id);
+        }
     }
 }
